fix: move size generator to Test project and add more generating methods
